requirements.txt
PyYAML==5.3.1
jsonschema==3.2.0
boto3==1.16.3
